@@ -1,13 +1,13 @@
 {
   "name": "vezgo-sdk-js",
-  "version": "1.1.13",
+  "version": "2.0.0",
   "description": "Official Vezgo JS SDK for the Browser & NodeJS",
   "author": "Wealthica Financial Technology Inc. <hello@wealthica.com> (https://wealthica.com/)",
   "license": "MIT",
-  "main": "lib/vezgo.js",
-  "types": "lib/vezgo.d.ts",
+  "main": "dist/vezgo.cjs.js",
+  "module": "dist/vezgo.es.js",
+  "types": "dist/index.d.ts",
   "files": [
-    "dist",
     "lib"
   ],
   "keywords": [
@@ -25,37 +25,36 @@
     "api"
   ],
   "scripts": {
-    "prebuild": "rimraf dist lib",
-    "build": "webpack",
+    "build": "rm -rf dist && yarn build:node && yarn build:browser",
+    "build:browser": "vite build --config vite.config.browser.mjs",
+    "build:node": "vite build --config vite.config.node.mjs",
     "prepublishOnly": "npm run build",
-    "test": "jest",
-    "watch:test": "jest --watch",
+    "test": "vitest run",
+    "watch:test": "vitest --watch",
     "lint": "eslint .",
     "lint-fix": "eslint --fix ."
   },
   "dependencies": {
+    "@rollup/plugin-commonjs": "^28.0.2",
     "apisauce": "^3.0.1",
-    "core-js": "^3.17.3",
-    "jsonwebtoken": "^9.0.2",
-    "regenerator-runtime": "^0.13.9"
+    "jwt-decode": "^4.0.0"
   },
   "devDependencies": {
-    "@babel/cli": "^7.14.8",
-    "@babel/core": "^7.15.0",
-    "@babel/preset-env": "^7.15.0",
-    "@babel/register": "^7.15.3",
-    "axios-mock-adapter": "^1.20.0",
-    "babel-loader": "^8.2.2",
-    "copy-webpack-plugin": "^10.2.4",
-    "eslint": "^7.32.0",
-    "eslint-config-airbnb-base": "^14.2.1",
-    "eslint-plugin-import": "^2.24.2",
-    "eslint-webpack-plugin": "^3.0.1",
-    "jest": "^29.7.0",
-    "node-polyfill-webpack-plugin": "^1.1.4",
-    "rimraf": "2.6.2",
-    "webpack": "~5.74.0",
-    "webpack-cli": "~4.10.0",
-    "webpack-merge": "^5.8.0"
+    "@babel/cli": "^7.24.6",
+    "@babel/core": "^7.24.6",
+    "@babel/preset-env": "^7.24.6",
+    "@babel/register": "^7.24.6",
+    "axios-mock-adapter": "^1.22.0",
+    "babel-loader": "^9.1.3",
+    "eslint": "^9.3.0",
+    "eslint-config-airbnb-base": "^15.0.0",
+    "eslint-plugin-import": "^2.29.1",
+    "jsonwebtoken": "^8.5.1",
+    "vite": "^6.2.0",
+    "vite-plugin-commonjs": "^0.10.1",
+    "vite-plugin-node-polyfills": "^0.23.0",
+    "vite-plugin-static-copy": "^1.0.5",
+    "vitest": "^1.6.0",
+    "prettier": "^2.7.1"
   }
 }
