@@ -2,7 +2,8 @@ require('dotenv').config();
 const express = require('express');
 const path = require('path');
 // For local development replace vezgo-sdk-js vezgo ../../lib/vezgo
-// const Vezgo = require('../../lib/vezgo');
+// const Vezgo = require('../../dist/vezgo.cjs');
+
 const Vezgo = require('vezgo-sdk-js');
 
 const app = express();
@@ -30,9 +31,10 @@ app.get('/assets/config.js', (req, res) => {
 
 app.get('/assets/vezgo.js', (req, res) => {
   res.setHeader('Content-Type', 'application/javascript');
-  // For local development replace vezgo-sdk-js vezgo ../../dist/vezgo.js
-  // res.sendFile(path.join(__dirname, '../../dist/vezgo.js'));
-  res.sendFile(path.join(__dirname, '/node_modules/vezgo-sdk-js/dist/vezgo.js'));
+  // For local development replace vezgo-sdk-js vezgo ../../lib/vezgo.js
+  // res.sendFile(path.join(__dirname, '../../dist/vezgo.umd.js'));
+
+  res.sendFile(path.join(__dirname, '/node_modules/vezgo-sdk-js/dist/vezgo.umd'));
 });
 
 app.get('/renderjson.js', (req, res) => {
